+++ packages/widgetbook/test/src/providers/injected_theme_state_test.dart
@@ -1,19 +1,19 @@
+import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:widgetbook/src/providers/device_state.dart';
-import 'package:widgetbook_models/widgetbook_models.dart';
+import 'package:widgetbook/src/providers/injected_theme_state.dart';
 
 import '../../helper/model_helper.dart';
 
 void main() {
   group(
-    '$DeviceState',
+    '$InjectedThemeState',
     () {
       test(
         'returns true when instance is the same',
         () {
-          final instance = DeviceState(
-            availableDevices: [],
-            currentDevice: Apple.iPhone11,
+          final instance = InjectedThemeState(
+            lightTheme: ThemeData(),
+            darkTheme: ThemeData(),
           );
 
           expect(
@@ -23,20 +23,59 @@ void main() {
         },
       );
 
+      group('.isLightThemeDefined returns', () {
+        test('true when theme is defined', () {
+          final instance = InjectedThemeState(lightTheme: ThemeData());
+          expect(instance.isLightThemeDefined, isTrue);
+        });
+
+        test('false when theme is defined', () {
+          final instance = InjectedThemeState();
+          expect(instance.isLightThemeDefined, isFalse);
+        });
+      });
+
+      group('.isDarkThemeDefined returns', () {
+        test('true when theme is defined', () {
+          final instance = InjectedThemeState(darkTheme: ThemeData());
+          expect(instance.isDarkThemeDefined, isTrue);
+        });
+
+        test('false when theme is defined', () {
+          final instance = InjectedThemeState();
+          expect(instance.isDarkThemeDefined, isFalse);
+        });
+      });
+
+      group('.areBothThemesDefined returns', () {
+        test('true when theme is defined', () {
+          final instance = InjectedThemeState(
+            darkTheme: ThemeData(),
+            lightTheme: ThemeData(),
+          );
+          expect(instance.areBothThemesDefined, isTrue);
+        });
+
+        test('false when theme is defined', () {
+          final instance = InjectedThemeState();
+          expect(instance.areBothThemesDefined, isFalse);
+        });
+      });
+
       group(
         'returns true when',
         () {
           test(
             'two instances with the same values are compared',
             () {
-              final instance1 = DeviceState(
-                availableDevices: [],
-                currentDevice: Apple.iPhone11,
+              final instance1 = InjectedThemeState(
+                lightTheme: ThemeData(),
+                darkTheme: ThemeData(),
               );
 
-              final instance2 = DeviceState(
-                availableDevices: [],
-                currentDevice: Apple.iPhone11,
+              final instance2 = InjectedThemeState(
+                lightTheme: ThemeData(),
+                darkTheme: ThemeData(),
               );
 
               expect(
@@ -49,15 +88,14 @@ void main() {
           test(
             'the hashCodes of two instances with the same values are compared',
             () {
-              final list = <Device>[];
-              final instance1 = DeviceState(
-                availableDevices: list,
-                currentDevice: Apple.iPhone11,
+              final instance1 = InjectedThemeState(
+                lightTheme: ThemeData(),
+                darkTheme: ThemeData(),
               );
 
-              final instance2 = DeviceState(
-                availableDevices: list,
-                currentDevice: Apple.iPhone11,
+              final instance2 = InjectedThemeState(
+                lightTheme: ThemeData(),
+                darkTheme: ThemeData(),
               );
 
               expectEqualHashCodes(instance1, instance2);